file architecture refactoring

+++ include/const.h
@@ -1,1 +1,1 @@
-typedef float type_t;
+typedef double type_t;
